+++ packages/core/src/components/selectable/selectable.interfaces.ts
@@ -1,4 +1,6 @@
 export interface SelectableChangeEvent {
   originalEvent: Event;
   checked: boolean;
+  name?: string;
+  value?: string;
 }