+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 
 ### Changed
 * **BREAKING** Banner: Vervang `icon-only` Button door Icon Button ([#3554](https://github.com/dso-toolkit/dso-toolkit/issues/3554))
+* Selectable: `value` optioneel en SelectableChangeEvent uitbreidingen ([#3689](https://github.com/dso-toolkit/dso-toolkit/issues/3689))
 
 ### Fixed
 * Mark Bar: Uitspreken van `current` en `totalCount` bij verandering ([#3660](https://github.com/dso-toolkit/dso-toolkit/issues/3660))